Create README.md
The purpose was to learn Calculus on a python based server. The answers were submitted to a platform named mobius that uses python syntax (exception * symbol is replaced by ^). Due to university guidlines, the question papers were not shared however many case studies can still be found with the lessons to sympy and introduction to matplotlib.